# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Import/Export management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /colis:
    get:
      operationId: listColis
      tags: [colis]
      summary: List all packages
      parameters:
        - name: arrivageId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [individuel, mixte]
        - name: dateDebut
          in: query
          schema:
            type: string
            format: date
        - name: dateFin
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColisDetail"
    post:
      operationId: createColis
      tags: [colis]
      summary: Create a new package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColisRequest"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColisDetail"
  /colis/{id}:
    get:
      operationId: getColis
      tags: [colis]
      summary: Get package by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColisDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateColis
      tags: [colis]
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColisRequest"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColisDetail"
    delete:
      operationId: deleteColis
      tags: [colis]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /colis-proprietaires/{id}/paiements:
    post:
      operationId: enregistrerPaiement
      tags: [paiements]
      summary: Record a payment for a package owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaiementRequest"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paiement"
  /dettes:
    get:
      operationId: listDettes
      tags: [paiements]
      summary: List all outstanding debts
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: dateDebut
          in: query
          schema:
            type: string
            format: date
        - name: dateFin
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetteDetail"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Colis:
      type: object
      properties:
        id:
          type: integer
        arrivageId:
          type: integer
        codeColisReel:
          type: string
        codeColisComplet:
          type: string
        type:
          type: string
          enum: [individuel, mixte]
        poids:
          type: number
        montant:
          type: number
        dateCreation:
          type: string
          format: date-time
      required:
        - id
        - arrivageId
        - codeColisReel
        - codeColisComplet
        - type
        - poids
        - montant
        - dateCreation
    ColisProprietaire:
      type: object
      properties:
        id:
          type: integer
        colisId:
          type: integer
        clientId:
          type: integer
        clientNom:
          type: string
        clientTelephone:
          type: string
        poids:
          type: number
        montantDu:
          type: number
        montantPaye:
          type: number
        solde:
          type: number
      required:
        - id
        - colisId
        - clientId
        - clientNom
        - clientTelephone
        - poids
        - montantDu
        - montantPaye
        - solde
    ColisDetail:
      allOf:
        - $ref: "#/components/schemas/Colis"
        - type: object
          properties:
            proprietaires:
              type: array
              items:
                $ref: "#/components/schemas/ColisProprietaire"
            arrivageDate:
              type: string
              format: date
          required:
            - proprietaires
            - arrivageDate
    CreateColisProprietaireRequest:
      type: object
      properties:
        clientId:
          type: integer
        poids:
          type: number
        montantDu:
          type: number
        montantPaye:
          type: number
      required:
        - clientId
        - poids
        - montantDu
        - montantPaye
    CreateColisRequest:
      type: object
      properties:
        arrivageId:
          type: integer
        codeColisReel:
          type: string
        type:
          type: string
          enum: [individuel, mixte]
        poids:
          type: number
        montant:
          type: number
        proprietaires:
          type: array
          items:
            $ref: "#/components/schemas/CreateColisProprietaireRequest"
      required:
        - arrivageId
        - codeColisReel
        - type
        - poids
        - montant
        - proprietaires
    Paiement:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        colisProprietaireId:
          type: integer
        montant:
          type: number
        datePaiement:
          type: string
          format: date-time
        mode:
          type: string
      required:
        - id
        - clientId
        - colisProprietaireId
        - montant
        - datePaiement
        - mode
    CreatePaiementRequest:
      type: object
      properties:
        montant:
          type: number
        mode:
          type: string
          enum: [especes, virement, cheque, mobile]
      required:
        - montant
        - mode
...
    DetteDetail:
      type: object
      properties:
        colisProprietaireId:
          type: integer
        clientId:
          type: integer
        clientNom:
          type: string
        clientTelephone:
          type: string
        colisCode:
          type: string
        arrivageDate:
          type: string
          format: date
        montantDu:
          type: number
        montantPaye:
          type: number
        solde:
          type: number
      required:
        - colisProprietaireId
        - clientId
        - clientNom
        - clientTelephone
        - colisCode
        - arrivageDate
        - montantDu
        - montantPaye
        - solde
...