# lib/api-zod/src/generated/types/createColisProprietaireRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateColisProprietaireRequest {
  clientId: number;
  poids: number;
  montantDu: number;
  montantPaye: number;
}