# lib/api-zod/src/generated/types/createColisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateColisProprietaireRequest } from "./createColisProprietaireRequest";
import type { CreateColisRequestType } from "./createColisRequestType";

export interface CreateColisRequest {
  arrivageId: number;
  codeColisReel: string;
  type: CreateColisRequestType;
  poids: number;
  montant: number;
  proprietaires: CreateColisProprietaireRequest[];
}