# lib/api-zod/src/generated/types/paiement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */

export interface Paiement {
  id: number;
  clientId: number;
  colisProprietaireId: number;
  montant: number;
  datePaiement: Date;
  mode: string;
}